Implement Andriy Plokhotnyuk suggestion. See https://twitter.com/aplokhotnyuk/status/1528755201636810753?s=20&t=1iBMOo2GdqFHfDGPrzy1Lg
